+++ Natom.Gestion.WebApp.Clientes.Backend.Entities/Model/Results/spReportOrdenDePedidoResult.cs
@@ -39,7 +39,7 @@ namespace Natom.Gestion.WebApp.Clientes.Backend.Entities.Model.Results
 		public decimal PesoTotalEnKilogramos { get; set; }
 		public string DetalleCodigo { get; set; }
 		public string DetalleDescripcion { get; set; }
-		public int DetalleCantidad { get; set; }
+		public decimal DetalleCantidad { get; set; }
 		public decimal DetallePesoUnitarioEnKilogramos { get; set; }
 		public decimal? DetallePrecioUnitario { get; set; }
 		public decimal DetallePesoTotalEnKilogramos { get; set; }
