+++ Natom.Gestion.WebApp.Clientes.Backend/Controllers/StockController.cs
@@ -112,7 +112,7 @@ namespace Natom.Gestion.WebApp.Clientes.Backend.Controllers
                 var manager = new StockManager(_serviceProvider);
                 var stockActual = await manager.ObtenerStockActualAsync(productoId, depositoId);
 
-                return Ok(new ApiResultDTO<int>
+                return Ok(new ApiResultDTO<decimal>
                 {
                     Success = true,
                     Data = stockActual
